requirements.txt
Django==3.1.6
djangorestframework==3.12.2
django-environ==0.4.5
djangorestframework-simplejwt==4.6.0
Pillow==8.1.0
